+++ src/plan/src/resolver/mod.rs
@@ -19,11 +19,11 @@ mod tree;
 pub struct PlanResolver<'a> {
     ctx: &'a SessionContext,
     config: Arc<PlanConfig>,
-    cache_key: u64,
+    cache_key: u128,
 }
 
 impl<'a> PlanResolver<'a> {
-    pub fn new(ctx: &'a SessionContext, config: Arc<PlanConfig>, cache_key: u64) -> Self {
+    pub fn new(ctx: &'a SessionContext, config: Arc<PlanConfig>, cache_key: u128) -> Self {
         Self {
             ctx,
             config,
